Remove trivial generator
