+++ lib/murker.rb
@@ -1,7 +1,6 @@
 require 'murker/version'
 require 'murker/spy'
 require 'murker/interaction'
-require 'murker/open_api_generator'
 
 module Murker
   def self.capture(&block)
