+++ lib/murker/validator.rb
@@ -3,8 +3,8 @@ require 'murker/generator'
 module Murker
   # Validates an interaction against a schema
   class Validator
-    def self.call(interaction:, schema:, generator_class: Generator)
-      interaction_schema = generator_class.call(interaction: interaction)
+    def self.call(interaction:, schema:)
+      interaction_schema = Generator.call(interaction: interaction)
 
       YAML.load(interaction_schema) == YAML.load(schema)
     end
